requirements.txt
groq
python-dotenv
streamlit
PyPDF2  # for PDF processing